init.js to nuxt plugin
